Refactor chip class and move independent chips into separate classes (#101507)
